Claude Code assisted modification: ECDH-ES+AxxxKW implementation + warning elimination

+++ src/concatkdf.c
@@ -142,7 +142,7 @@ uint8_t *cjose_concatkdf_derive(const size_t keylen,
     }
 
     size_t offset = 0, amt = keylen;
-    for (int idx = 1; N >= idx; idx++)
+    for (size_t idx = 1; N >= idx; idx++)
     {
         uint8_t counter[4];
         _apply_uint32(idx, counter);
